# Add SharedPerformanceTest and improve Pool performance.  Still not good though.