Pool is still a lot worse than std::vector<std::unique_ptr<T>> for both insertion and iteration.  The goal was at least to be better than it for iteration due to the objects being sequential in memory.

I probably need to consider a way to just use vector<T> or vector<unique_ptr<T>>  I am not even sure it is a good idea to have the engine control all the allocation.  Maybe having vectors in the game code would be better.

[ RUN      ] PoolTests.SmallObject_InsertionPerformance
Pool<SmallObject>: 81089 us
dtor tryToFindMemoryTime = 6489 us
std::vector<SmallObject>: 707 us
std::vector<std::unique_ptr<SmallObject>>: 2927 us
[       OK ] PoolTests.SmallObject_InsertionPerformance (96 ms)
[ RUN      ] PoolTests.SmallObject_IterationPerformance
Pool<SmallObject>: 808 us
dtor tryToFindMemoryTime = 2110992554584 us
std::vector<SmallObject>: 152 us
std::vector<std::unique_ptr<SmallObject>>: 386 us